Fix parsing for multi-line summaries (#18)
Fix parsing for multi-line summaries. Summary and description are
separated by a blank newline. Maintain formatting of the description if
there are multiple sections separated by blank newlines.

Will require further changes in order to treat the description in this
edge case as `PyFunc.description` instead of `PyFunc.summary`:
```
"""

This is the description.

Args:
    x: The first argument.
    y: The second argument.

Returns:
    The sum of x and y.
"""
```