Skip first H1 in search index to avoid duplicates (#61)
Most markdown pages start with an H1 that repeats the page title. The search index was creating two entries for these pages: one for the page itself, and one for that first heading. This means when you search for a page, you get the same result twice.

This PR addresses the issue by not indexing the first H1 on a page.